refactor: extracted helper methods to own file
